mtt/requirements.txt
Django==4.2.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
wait-for-it==2.2.1
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-vite==2.1.3
django-filter==23.3
openpyxl==3.1.2
